Fix voted concern
