+++ app/controllers/concerns/voted.rb
@@ -1,4 +1,4 @@
-module Votable
+module Voted
   extend ActiveSupport::Concern
 
   included do
@@ -28,5 +28,6 @@ module Votable
 
   def set_vote
     @votable = model_klass.find(params[:id])
+    instance_variable_set("@#{controller_name.singularize}", @votable)
   end
 end
