+++ app/models/concerns/has_votable.rb
@@ -1,0 +1,6 @@
+module HasVotable
+  extend ActiveSupport::Concern
+  included do
+    has_many :votes, as: :votable, dependent: :destroy
+  end
+end
