+++ app/models/question.rb
@@ -2,10 +2,9 @@ class Question < ActiveRecord::Base
 
   include Attachmentable
   include HasUser
-  has_many :votes, as: :votable, dependent: :destroy
+  include HasVotable
 
   has_many :answers, dependent: :destroy
 
   validates :title, :body, presence: true
-
 end
