+++ app/views/questions/show.html.slim
@@ -1,5 +1,5 @@
 .row
-  .col-md-6.question__wrapper id="#{dom_id(@question)}"
+  .col-md-6.question_wrapper id="#{dom_id(@question)}"
     h2 = @question.title
     p = @question.body
     - if @question.attachments.any?
